requirements.txt
tqdm
scipy
pillow
matplotlib
scikit-learn
numpy
omegaconf
ftfy
einops
termcolor

xlrd==1.2.0
torch==1.9.0
torchvision==0.10.0
torchaudio==0.9.0
